+++ package.json
@@ -12,7 +12,9 @@
   "dependencies": {
     "copilotkit": "^0.0.12",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^7.1.5",
+    "react-toastify": "^11.0.3"
   },
   "devDependencies": {
     "@eslint/js": "^9.17.0",
