+++ src/App.css
@@ -1,239 +1,35 @@
-.ai-assistant-container {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  grid-gap: 20px;
-  align-items: flex-start;
-  padding: 20px;
-  background-color: #f4f7fb;
-  border-radius: 15px;
-  max-width: 1000px;
-  margin: 20px auto;
-  box-shadow: 0 4px 15px rgba(0, 0, 0, 0.1);
-}
-
-.text-section, .summary-section {
-  background-color: #fff;
-  border-radius: 10px;
-  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
-  padding: 20px;
-}
-
-.text-section {
-  background-color: #e8f4fd;
-}
-
-.summary-section {
-  background-color: #fff3e6;
-}
-
-h2 {
-  color: #333;
-  font-size: 1.75rem;
-  font-weight: bold;
-  margin-bottom: 15px;
-  text-align: center;
-}
-
-textarea {
-  width: 100%;
-  padding: 12px;
-  font-size: 1.1rem;
-  font-family: 'Arial', sans-serif;
-  border: 1px solid #ccc;
-  border-radius: 8px;
-  box-sizing: border-box;
-  background-color: #fff;
-  resize: vertical;
-  min-height: 150px;
-}
-
-textarea:focus {
-  outline: none;
-  border-color: #3b9dd9;
-  box-shadow: 0 0 6px rgba(59, 157, 217, 0.5);
-}
-
-button {
-  background-color: #3b9dd9;
-  color: white;
-  border: none;
-  padding: 12px 18px;
-  font-size: 1.1rem;
-  border-radius: 8px;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-  width: 100%;
-}
-
-button:hover {
-  background-color: #1f8dc6;
-}
-
-button:focus {
-  outline: none;
-}
-
-.ai-suggestion p {
-  color: #777;
-  font-size: 0.95rem;
-  margin-top: 8px;
-}
-
-textarea[placeholder] {
-  color: #aaa;
-  font-style: italic;
-}
-
-textarea[readonly] {
-  background-color: #f5f5f5;
-}
-
-.suggestions-list {
-  margin-top: 10px;
-  padding: 0;
-  max-height: 180px;
-  overflow-y: auto;
-  background-color: #f9f9f9;
-  border-radius: 8px;
-}
-
-.suggestions-list ul {
-  list-style-type: none;
-  margin: 0;
-  padding: 0;
-}
-
-.suggestions-list li {
-  padding: 8px;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-  border-bottom: 1px solid #e0e0e0;
-}
-
-.suggestions-list li:hover {
-  background-color: #e3f1fd;
-}
-
-/* Mobile Responsive */
-@media (max-width: 768px) {
-  .ai-assistant-container {
-    grid-template-columns: 1fr;
-    padding: 15px;
-  }
-
-  .text-section, .summary-section {
-    width: 100%;
-    margin-bottom: 20px;
-  }
-
-  h2 {
-    font-size: 1.5rem;
-  }
-
-  textarea {
-    font-size: 1rem;
-  }
-
-  button {
-    font-size: 1rem;
-  }
-}
-
-@media (max-width: 480px) {
-  .ai-assistant-container {
-    padding: 10px;
-  }
-
-  h2 {
-    font-size: 1.25rem;
-  }
-
-  textarea {
-    font-size: 0.95rem;
-  }
-
-  button {
-    font-size: 0.95rem;
-  }
-}
-
-.employee-ticket-section {
-  margin-top: 30px;
-  background-color: #fff;
-  border-radius: 10px;
-  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
-  padding: 20px;
-}
-
-.employee-ticket-section h2 {
-  text-align: center;
-  margin-bottom: 20px;
-}
-
-table {
-  width: 100%;
-  border-collapse: collapse;
-}
-
-table th,
-table td {
-  padding: 10px;
-  text-align: left;
-  border-bottom: 1px solid #ddd;
-}
-
-table th {
-  background-color: #f4f7fb;
-}
-
-table tr:hover {
-  background-color: #f9f9f9;
-}
-
-/* Modal Overlay */
-.modal-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
+/* Header */
+.header {
   display: flex;
   justify-content: center;
-  align-items: center;
-  z-index: 1000;
-}
-
-/* Modal Box */
-.modal {
-  background-color: white;
-  padding: 20px;
-  border-radius: 10px;
-  box-shadow: 0 4px 15px rgba(0, 0, 0, 0.1);
-  max-width: 500px;
-  width: 100%;
-}
-
-.modal h3 {
-  margin-top: 0;
-}
-
-.modal p {
-  font-size: 1rem;
-  margin: 10px 0;
-}
-
-.modal button {
   background-color: #3b9dd9;
+  padding: 15px;
+  border-radius: 10px;
+  margin-bottom: 20px;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+}
+
+.nav-links {
+  display: flex;
+  gap: 20px;
+}
+
+.nav-item {
   color: white;
-  border: none;
-  padding: 12px 18px;
-  font-size: 1rem;
+  text-decoration: none;
+  font-size: 1.1rem;
+  padding: 10px 15px;
   border-radius: 8px;
-  cursor: pointer;
   transition: background-color 0.3s ease;
 }
 
-.modal button:hover {
-  background-color: #1f8dc6;
+.nav-item:hover {
+  background-color: rgba(255, 255, 255, 0.2);
+}
+
+/* Active Page Highlight */
+.nav-item.active {
+  background-color: #fff;
+  color: #3b9dd9;
+  font-weight: 600;
 }
