+++ src/components/AIAssistant.css
@@ -1,0 +1,169 @@
+/* AIAssistant.css */
+
+.ai-assistant-container {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    gap: 20px;
+    align-items: flex-start;
+    padding: 25px;
+    background-color: #eef5ff;
+    border-radius: 15px;
+    max-width: 1000px;
+    margin: 20px auto;
+    box-shadow: 0 6px 15px rgba(0, 0, 0, 0.12);
+  }
+  
+  /* Left and Right Sections */
+  .text-section, 
+  .summary-section {
+    background-color: #fff;
+    border-radius: 12px;
+    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.12);
+    padding: 25px;
+    transition: transform 0.2s ease-in-out;
+  }
+  
+  .text-section:hover,
+  .summary-section:hover {
+    transform: translateY(-5px);
+  }
+  
+  .text-section {
+    background-color: #e3f2fd;
+  }
+  
+  .summary-section {
+    background-color: #fffaf0;
+  }
+  
+  /* Headings */
+  h2 {
+    color: #222;
+    font-size: 1.8rem;
+    font-weight: bold;
+    margin-bottom: 18px;
+    text-align: center;
+  }
+  
+  /* Textarea */
+  textarea {
+    width: 100%;
+    padding: 14px;
+    font-size: 1.1rem;
+    font-family: 'Arial', sans-serif;
+    border: 1px solid #ccc;
+    border-radius: 10px;
+    box-sizing: border-box;
+    background-color: #fff;
+    resize: vertical;
+    min-height: 160px;
+    transition: border-color 0.3s ease-in-out, box-shadow 0.3s ease-in-out;
+  }
+  
+  textarea:focus {
+    outline: none;
+    border-color: #3498db;
+    box-shadow: 0 0 6px rgba(52, 152, 219, 0.5);
+  }
+  
+  /* Button */
+  button {
+    background-color: #007bff;
+    color: white;
+    border: none;
+    padding: 14px 20px;
+    font-size: 1.1rem;
+    border-radius: 10px;
+    cursor: pointer;
+    transition: background-color 0.3s ease, transform 0.2s ease;
+    width: 100%;
+  }
+  
+  button:hover {
+    background-color: #0056b3;
+    transform: scale(1.05);
+  }
+  
+  button:focus {
+    outline: none;
+  }
+  
+  /* AI Suggestions */
+  .ai-suggestion p {
+    color: #555;
+    font-size: 0.95rem;
+    margin-top: 10px;
+  }
+
+  
+  /* Suggestions List */
+  .suggestions-list {
+    margin-top: 12px;
+    max-height: 200px;
+    overflow-y: auto;
+    background-color: #f9f9f9;
+    border-radius: 10px;
+    padding: 10px;
+  }
+  
+  .suggestions-list ul {
+    list-style-type: none;
+    margin: 0;
+    padding: 0;
+  }
+  
+  .suggestions-list li {
+    padding: 10px;
+    cursor: pointer;
+    transition: background-color 0.3s ease;
+    border-bottom: 1px solid #ddd;
+  }
+  
+  .suggestions-list li:hover {
+    background-color: #d6eaff;
+  }
+  
+  /* Mobile Responsive */
+  @media (max-width: 768px) {
+    .ai-assistant-container {
+      grid-template-columns: 1fr;
+      padding: 20px;
+    }
+  
+    .text-section, 
+    .summary-section {
+      width: 100%;
+      margin-bottom: 20px;
+    }
+  
+    h2 {
+      font-size: 1.6rem;
+    }
+  
+    textarea {
+      font-size: 1rem;
+    }
+  
+    button {
+      font-size: 1rem;
+    }
+  }
+  
+  @media (max-width: 480px) {
+    .ai-assistant-container {
+      padding: 15px;
+    }
+  
+    h2 {
+      font-size: 1.4rem;
+    }
+  
+    textarea {
+      font-size: 0.95rem;
+    }
+  
+    button {
+      font-size: 0.95rem;
+    }
+  }
+
